# GODRIVER-2021 Use pooling to reduce memory allocation.