--- a/x/mongo/driver/operation/command.go
+++ b/x/mongo/driver/operation/command.go
@@ -106,7 +106,7 @@ func (c *Command) Execute(ctx context.Context) error {
 		Crypt:          c.crypt,
 		ServerAPI:      c.serverAPI,
 		Timeout:        c.timeout,
-	}.Execute(ctx, nil)
+	}.Execute(ctx)
 }
 
 // Session sets the session for this operation.