--- a/x/mongo/driver/operation/hello.go
+++ b/x/mongo/driver/operation/hello.go
@@ -194,7 +194,7 @@ func (h *Hello) Execute(ctx context.Context) error {
 		return errors.New("a Hello must have a Deployment set before Execute can be called")
 	}
 
-	return h.createOperation().Execute(ctx, nil)
+	return h.createOperation().Execute(ctx)
 }
 
 // StreamResponse gets the next streaming Hello response from the server.
@@ -229,7 +229,7 @@ func (h *Hello) GetHandshakeInformation(ctx context.Context, _ address.Address, 
 			return nil
 		},
 		ServerAPI: h.serverAPI,
-	}.Execute(ctx, nil)
+	}.Execute(ctx)
 	if err != nil {
 		return driver.HandshakeInformation{}, err
 	}