--- a/x/mongo/driver/operation_exhaust.go
+++ b/x/mongo/driver/operation_exhaust.go
@@ -19,7 +19,7 @@ func (op Operation) ExecuteExhaust(ctx context.Context, conn StreamerConnection,
 	}
 
 	scratch = scratch[:0]
-	res, err := op.readWireMessage(ctx, conn, scratch)
+	res, err := op.readWireMessage(ctx, conn, &scratch)
 	if err != nil {
 		return err
 	}