--- a/x/mongo/driver/operation_exhaust.go
+++ b/x/mongo/driver/operation_exhaust.go
@@ -13,13 +13,18 @@ import (
 
 // ExecuteExhaust reads a response from the provided StreamerConnection. This will error if the connection's
 // CurrentlyStreaming function returns false.
-func (op Operation) ExecuteExhaust(ctx context.Context, conn StreamerConnection, scratch []byte) error {
+func (op Operation) ExecuteExhaust(ctx context.Context, conn StreamerConnection) error {
 	if !conn.CurrentlyStreaming() {
 		return errors.New("exhaust read must be done with a connection that is currently streaming")
 	}
 
-	scratch = scratch[:0]
-	res, _, err := op.readWireMessage(ctx, conn, scratch)
+	wm := memoryPool.Get().(*[]byte)
+	defer func() {
+		memoryPool.Put(wm)
+	}()
+	var res []byte
+	var err error
+	res, *wm, err = op.readWireMessage(ctx, conn, (*wm)[:0])
 	if err != nil {
 		return err
 	}