[GODRIVER-2021](https://jira.mongodb.org/browse/GODRIVER-2021)
[GODRIVER-1670](https://jira.mongodb.org/browse/GODRIVER-1670)

Use pooling to reduce memory allocation.

The benchstat from benchmark/operation_test.go shows:
```
name                          old time/op    new time/op    delta
ClientWrite/not_compressed-4     153µs ± 2%     154µs ± 2%     ~     (p=0.079 n=9+10)
ClientWrite/snappy-4             183µs ± 2%     187µs ±12%     ~     (p=0.878 n=8+8)
ClientWrite/zlib-4              1.35ms ±36%    1.39ms ±27%     ~     (p=0.842 n=10+9)
ClientWrite/zstd-4               286µs ±13%     286µs ±16%     ~     (p=0.666 n=9+9)

name                          old alloc/op   new alloc/op   delta
ClientWrite/not_compressed-4    34.3kB ± 0%    23.9kB ± 0%  -30.32%  (p=0.000 n=10+9)
ClientWrite/snappy-4            52.1kB ± 0%    36.4kB ± 0%  -30.14%  (p=0.000 n=10+10)
ClientWrite/zlib-4               907kB ± 0%     896kB ± 0%   -1.26%  (p=0.000 n=10+8)
ClientWrite/zstd-4              54.1kB ± 0%    39.5kB ± 0%  -26.99%  (p=0.000 n=9+10)

name                          old allocs/op  new allocs/op  delta
ClientWrite/not_compressed-4      59.0 ± 0%      54.0 ± 0%   -8.47%  (p=0.000 n=10+10)
ClientWrite/snappy-4              66.0 ± 0%      57.0 ± 0%  -13.64%  (p=0.000 n=10+10)
ClientWrite/zlib-4                 103 ± 0%        96 ± 0%   -6.80%  (p=0.000 n=8+7)
ClientWrite/zstd-4                 108 ± 0%        99 ± 0%   -8.33%  (p=0.000 n=10+10)
```

```
name                         old time/op    new time/op    delta
ClientRead/not_compressed-4     248µs ± 2%     256µs ±11%     ~     (p=1.000 n=8+9)
ClientRead/snappy-4             270µs ± 1%     272µs ± 2%     ~     (p=0.113 n=9+10)
ClientRead/zlib-4              1.78ms ± 3%    1.78ms ± 2%     ~     (p=0.863 n=9+9)
ClientRead/zstd-4               399µs ± 2%     399µs ± 3%     ~     (p=0.971 n=10+10)

name                         old alloc/op   new alloc/op   delta
ClientRead/not_compressed-4    26.5kB ± 0%    26.1kB ± 0%   -1.59%  (p=0.000 n=9+10)
ClientRead/snappy-4            44.1kB ± 0%    43.6kB ± 0%   -1.08%  (p=0.000 n=10+10)
ClientRead/zlib-4               899kB ± 0%     903kB ± 0%   +0.48%  (p=0.000 n=9+10)
ClientRead/zstd-4               124kB ± 0%     123kB ± 0%   -1.11%  (p=0.000 n=10+10)

name                         old allocs/op  new allocs/op  delta
ClientRead/not_compressed-4      73.0 ± 0%      67.0 ± 0%   -8.22%  (p=0.000 n=10+10)
ClientRead/snappy-4              80.0 ± 0%      70.0 ± 0%  -12.50%  (p=0.000 n=10+10)
ClientRead/zlib-4                 130 ± 0%       122 ± 0%   -5.77%  (p=0.000 n=8+10)
ClientRead/zstd-4                 137 ± 0%       127 ± 0%   -7.30%  (p=0.000 n=10+10)
```

coauthor: Matt Dale