@qingyang-hu 2022-08-24
Added benchmark for bulk write. It shows:
```
name                              old time/op    new time/op    delta
ClientBulkWrite/not_compressed-4    3.19ms ±43%    3.20ms ±37%     ~     (p=0.436 n=9+9)
ClientBulkWrite/snappy-4            3.51ms ±26%    3.19ms ±38%     ~     (p=0.387 n=9+9)
ClientBulkWrite/zlib-4              3.93ms ±17%    3.87ms ± 7%     ~     (p=0.798 n=8+8)
ClientBulkWrite/zstd-4              3.43ms ±24%    3.56ms ±41%     ~     (p=0.546 n=9+9)

name                              old alloc/op   new alloc/op   delta
ClientBulkWrite/not_compressed-4     207kB ± 0%      90kB ± 1%  -56.57%  (p=0.000 n=10+9)
ClientBulkWrite/snappy-4             264kB ± 0%     146kB ± 2%  -44.63%  (p=0.000 n=10+10)
ClientBulkWrite/zlib-4              1.08MB ± 0%    1.02MB ± 1%   -6.21%  (p=0.000 n=9+10)
ClientBulkWrite/zstd-4               273kB ± 0%     152kB ± 0%  -44.29%  (p=0.000 n=10+9)

name                              old allocs/op  new allocs/op  delta
ClientBulkWrite/not_compressed-4       124 ± 0%       115 ± 1%   -7.53%  (p=0.000 n=9+9)
ClientBulkWrite/snappy-4               132 ± 0%       118 ± 1%  -10.04%  (p=0.000 n=10+10)
ClientBulkWrite/zlib-4                 170 ± 0%       161 ± 1%   -5.24%  (p=0.000 n=9+9)
ClientBulkWrite/zstd-4                 184 ± 0%       171 ± 0%   -7.07%  (p=0.000 n=10+8)
```