@qingyang-hu 2022-08-30
Also, dropped collections after benchmarking.